chore: remove dead Helm version code (#19449)
